Push remaining local changes

+++ apps/student/lib/face/camera_input_converter.dart
@@ -1,5 +1,4 @@
 import 'dart:io';
-import 'dart:ui';
 
 import 'package:camera/camera.dart';
 import 'package:flutter/services.dart';
@@ -39,7 +38,8 @@ class CameraInputConverter {
     }
     if (rotation == null) return null;
 
-    final format = InputImageFormatValue.fromRawValue(image.format.raw);
+    final format =
+        InputImageFormatValue.fromRawValue(image.format.raw as int);
     if (format == null ||
         (Platform.isAndroid && format != InputImageFormat.nv21) ||
         (Platform.isIOS && format != InputImageFormat.bgra8888)) {
